+++ AdonaPerfuma/Interfaces/IProductRepo.cs
@@ -8,12 +8,12 @@ namespace AdonaPerfuma.Interfaces
     public interface IProductRepo
     {
          Task<List<Product>> GetAllProducts();
-         Task<Product> GetProductById(int id);
-
+         Task<Product> GetProductById(long id);
+        
          Task AddProduct(Product product);
 
-         Task UpdateProduct(int id,Product modifiedProduct);
+         Task UpdateProduct(long id,Product modifiedProduct);
 
-         Task DeleteProduct(int id);
+         Task DeleteProduct(long id);
     }
 }
